Collapse Unsloth shared API into train module
